Added protection before remove object which contains any Item instance

+++ src/main/java/com/company/ordersbackend/service/ItemCategoryService.java
@@ -44,8 +44,11 @@ public class ItemCategoryService {
 
     public boolean deleteById(long id) {
         if (itemCategoryRepository.existsById(id)) {
-            itemCategoryRepository.deleteById(id);
-            return true;
+            if(!itemCategoryRepository.findById(id).get().getItemList().isEmpty()){
+                itemCategoryRepository.deleteById(id);
+                return true;
+            }
+
         }
 
         return false;
