+++ src/main/java/com/company/ordersbackend/service/ProducerService.java
@@ -45,8 +45,11 @@ public class ProducerService {
 
     public boolean deleteById(long id) {
         if (producerRepository.existsById(id)) {
-            producerRepository.deleteById(id);
-            return true;
+            if(!producerRepository.findById(id).get().getItemList().isEmpty()){
+                producerRepository.deleteById(id);
+                return true;
+            }
+
         }
         return false;
     }
